Make hand enum part of player
