+++ src/main/java/cam72cam/mod/config/ConfigGui.java
@@ -1,7 +1,7 @@
 package cam72cam.mod.config;
 
+import cam72cam.mod.entity.Player;
 import cam72cam.mod.gui.*;
-import cam72cam.mod.util.Hand;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiScreen;
 
@@ -28,7 +28,7 @@ public class ConfigGui implements IScreen {
             ci.read();
             widgets.add(screen -> new Button(screen, 0 - 100,  finalI * 20, 200, 20, type.getSimpleName()) {
                 @Override
-                public void onClick(Hand hand) {
+                public void onClick(Player.Hand hand) {
                     Minecraft.getMinecraft().displayGuiScreen(new ScreenBuilder(new ConfigGui(ConfigGui.this, ci.pc, ci)));
                 }
             });
@@ -55,7 +55,7 @@ public class ConfigGui implements IScreen {
                 widgets.add(screen -> {
                     Button btn = new Button(screen, 0 - 100, offsetI * 20, 200, 20, property.getName()) {
                         @Override
-                        public void onClick(Hand hand) {
+                        public void onClick(Player.Hand hand) {
                             Minecraft.getMinecraft().displayGuiScreen(new ScreenBuilder(new ConfigGui(ConfigGui.this, (ConfigFile.PropertyClass) property, ci)));
                         }
                     };
@@ -88,7 +88,7 @@ public class ConfigGui implements IScreen {
                         widgets.add(screen -> {
                             Button btn = new Button(screen, -1, offsetI * 20, 200, 20, val.toString()) {
                                 @Override
-                                public void onClick(Hand hand) {
+                                public void onClick(Player.Hand hand) {
                                     Boolean value = !Boolean.parseBoolean(this.getText());
                                     this.setText(value.toString());
                                     try {
@@ -109,7 +109,7 @@ public class ConfigGui implements IScreen {
                             Button btn = new Button(screen, -1, offsetI * 20, 200, 20, val.toString()) {
                                 Enum curr = val;
                                 @Override
-                                public void onClick(Hand hand) {
+                                public void onClick(Player.Hand hand) {
                                     curr = arry[(curr.ordinal()+1) % arry.length];
                                     this.setText(curr.toString());
                                     try {
@@ -184,7 +184,7 @@ public class ConfigGui implements IScreen {
                 widgets.add(screen -> {
                     Button btn = new Button(screen, -200, offsetI * 20, 200, 20, property.getName()) {
                         @Override
-                        public void onClick(Hand hand) {
+                        public void onClick(Player.Hand hand) {
 
                         }
                     };
@@ -202,7 +202,7 @@ public class ConfigGui implements IScreen {
             int i = 0;
 
             @Override
-            public void onClick(Hand hand) {
+            public void onClick(Player.Hand hand) {
                 i++;
                 if (i >= pages) {
                     i = 0;
