+++ src/main/java/cam72cam/mod/event/ClientEvents.java
@@ -2,11 +2,11 @@ package cam72cam.mod.event;
 
 import cam72cam.mod.ModCore;
 import cam72cam.mod.entity.EntityRegistry;
+import cam72cam.mod.entity.Player;
 import cam72cam.mod.input.Mouse;
 import cam72cam.mod.render.EntityRenderer;
 import cam72cam.mod.render.GlobalRender;
 import cam72cam.mod.sound.Audio;
-import cam72cam.mod.util.Hand;
 import net.minecraft.client.Minecraft;
 import net.minecraftforge.client.event.*;
 import net.minecraftforge.client.event.sound.SoundLoadEvent;
@@ -37,7 +37,7 @@ public class ClientEvents {
     }
 
     public static final Event<Runnable> TICK = new Event<>();
-    public static final Event<Function<Hand, Boolean>> CLICK = new Event<>();
+    public static final Event<Function<Player.Hand, Boolean>> CLICK = new Event<>();
     public static final Event<Runnable> MODEL_CREATE = new Event<>();
     public static final Event<Consumer<ModelBakeEvent>> MODEL_BAKE = new Event<>();
     public static final Event<Runnable> TEXTURE_STITCH = new Event<>();
@@ -65,7 +65,7 @@ public class ClientEvents {
             int useID = Minecraft.getMinecraft().gameSettings.keyBindUseItem.getKeyCode() + 100;
 
             if ((event.getButton() == attackID || event.getButton() == useID) && event.isButtonstate()) {
-                Hand button = attackID == event.getButton() ? Hand.SECONDARY : Hand.PRIMARY;
+                Player.Hand button = attackID == event.getButton() ? Player.Hand.SECONDARY : Player.Hand.PRIMARY;
                 if (!CLICK.executeCancellable(x -> x.apply(button))) {
                     event.setCanceled(true);
                 }
